feat(editor): Show Slack docs link for agent setup (no-changelog)
